Split out the class ChatFullInfo from the class Chat and changed the return type of the method getChat to ChatFullInfo.
Note: this is a breaking change that may require updating the user code
